+++ app/Providers/AppServiceProvider.php
@@ -69,7 +69,7 @@ class AppServiceProvider extends ServiceProvider
         Vite::useHotFile(storage_path('vite.hot')) // Customize the "hot" file...
             ->useBuildDirectory('storage/vite') // Customize the build directory...
             ->withEntryPoints([
-                'resources/js/app.js'
+                'resources/js/classic.js'
             ]);