+++ resources/js/bootstrap.js
@@ -11,4 +11,8 @@ window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 window.axios.defaults.headers.common['X-CSRF-TOKEN'] = document.head.querySelector('meta[name="csrf-token"]').content;
 axios.defaults.withCredentials = true;
 
-import './echo'
+if (import.meta.env.VITE_REVERB_APP_KEY) {
+    import('./echo');
+} else {
+    window.Echo = null;
+}