+++ resources/views/layouts/app.blade.php
@@ -9,7 +9,7 @@
     @include('layouts.shared/title-meta', ['title' => $title ?? null])
     @yield('css')
     @include('layouts.shared/head-css', ['mode' => $mode ?? '', 'demo' => $demo ?? ''])
-    @vite(['resources/js/app.js','resources/js/hyper-head.js', 'resources/js/hyper-config.js'])
+    @vite(['resources/js/classic.js','resources/js/hyper-head.js', 'resources/js/hyper-config.js'])
 
     @stack('head.end')
 </head>