+++ resources/views/layouts/shared/footer-scripts.blade.php
@@ -1,10 +1,9 @@
 <!-- bundle -->
+@vite(['resources/js/ui/component.toastr.js'])
+
 @yield('script')
 @yield('script-bottom')
 @stack('scripts')
-
-@vite(['node_modules/select2/dist/css/select2.min.css', 'resources/js/ui/component.toastr.js', 'node_modules/jquery-toast-plugin/dist/jquery.toast.min.css'])
-
 
 <script>
     document.addEventListener('DOMContentLoaded', function() {